requirements.txt
streamlit>=1.35
openpyxl>=3.1
pandas>=2.0
anthropic>=0.25
matplotlib>=3.8
reportlab>=4.0
python-dotenv>=1.0
xlsxwriter>=3.1
pypdf>=4.0
python-docx>=1.1
Pillow>=10.0
numpy>=1.24
